+++ src/FSharpVSPowerTools.Core/DepthParser.fs
@@ -329,7 +329,7 @@ open DepthParsing
 open System
 
 type DepthParser private () =
-    let checker = lazy (InteractiveChecker.Create())
+    let checker = lazy (FSharpChecker.Create())
 
     member internal x.Checker = checker.Value
 
@@ -339,7 +339,7 @@ type DepthParser private () =
 
     // TODO, consider perf, way just to consider viewport and do as little work as necessary?
 
-    member internal x.GetRangesImpl(sourceCodeLinesOfTheFile, sourceCodeOfTheFile, filename, checker: InteractiveChecker option) =
+    member internal x.GetRangesImpl(sourceCodeLinesOfTheFile, sourceCodeOfTheFile, filename, checker: FSharpChecker option) =
         async {
             let mindents = computeMinIndentArray sourceCodeLinesOfTheFile
             let indent startLine endLine =
@@ -371,7 +371,7 @@ type DepthParser private () =
     /// Get the "nested ranges" of source code structures, along with the minimum-number-of-indent spaces inside that span (ignoring comments and #commands).
     /// Note: The 'filename' is only used e.g. to look at the filename extension (e.g. ".fs" versus ".fsi"), this does not try to load the file off disk.  
     ///       Instead, 'sourceCodeOfTheFile' should contain the entire file as a giant string.
-    static member GetRanges(sourceCodeOfTheFile: string, filename, ?checker: InteractiveChecker) =
+    static member GetRanges(sourceCodeOfTheFile: string, filename, ?checker: FSharpChecker) =
         let sourceCodeLinesOfTheFile = sourceCodeOfTheFile.Split [|'\n'|]
         DepthParser.Instance.GetRangesImpl(sourceCodeLinesOfTheFile, sourceCodeOfTheFile, filename, checker)
 
@@ -384,7 +384,7 @@ type DepthParser private () =
     /// Get non-overlapping ranges, where each range spans at most a single line, and has info about its "semantic depth".
     /// Note: The 'filename' is only used e.g. to look at the filename extension (e.g. ".fs" versus ".fsi"), this does not try to load the file off disk.  
     ///       Instead, 'sourceCodeOfTheFile' should contain the entire file as a giant string.
-    static member GetNonoverlappingDepthRanges(sourceCodeOfTheFile: string, filename, ?checker: InteractiveChecker) =
+    static member GetNonoverlappingDepthRanges(sourceCodeOfTheFile: string, filename, ?checker: FSharpChecker) =
         async {
             let sourceCodeLinesOfTheFile = sourceCodeOfTheFile.Split([|"\r\n";"\n"|], StringSplitOptions.None)
             let lineLens = sourceCodeLinesOfTheFile |> Seq.map (fun s -> s.TrimEnd(null).Length) |> (fun s -> Seq.append s [0]) |> Seq.toArray