+++ src/FSharpVSPowerTools.Core/InterfaceStubGenerator.fs
@@ -126,7 +126,7 @@ module InterfaceStubGenerator =
         else argName),
         namesWithIndices
 
-    let internal formatArgsUsage ctx hasTypeAnnotation (v: FSharpMemberFunctionOrValue) args =
+    let internal formatArgsUsage ctx hasTypeAnnotation (v: FSharpMemberOrFunctionOrValue) args =
         let isItemIndexer = (v.IsInstanceMember && v.DisplayName = "Item")
         let unit, argSep, tupSep = "()", " ", ", "
         let args, namesWithIndices =
@@ -152,10 +152,10 @@ module InterfaceStubGenerator =
 
     [<RequireQualifiedAccess; NoComparison>]
     type internal MemberInfo =
-        | PropertyGetSet of FSharpMemberFunctionOrValue * FSharpMemberFunctionOrValue
-        | Member of FSharpMemberFunctionOrValue
+        | PropertyGetSet of FSharpMemberOrFunctionOrValue * FSharpMemberOrFunctionOrValue
+        | Member of FSharpMemberOrFunctionOrValue
 
-    let internal getArgTypes (ctx: Context) (v: FSharpMemberFunctionOrValue) =
+    let internal getArgTypes (ctx: Context) (v: FSharpMemberOrFunctionOrValue) =
         let argInfos = v.CurriedParameterGroups |> Seq.map Seq.toList |> Seq.toList 
             
         let retType = v.ReturnParameter.Type
@@ -186,18 +186,18 @@ module InterfaceStubGenerator =
         argInfos, retType
 
     /// Convert a getter/setter to its canonical form
-    let internal normalizePropertyName (v: FSharpMemberFunctionOrValue) =
+    let internal normalizePropertyName (v: FSharpMemberOrFunctionOrValue) =
         let displayName = v.DisplayName
         if (v.IsPropertyGetterMethod && displayName.StartsWith("get_")) || 
             (v.IsPropertySetterMethod && displayName.StartsWith("set_")) then
             displayName.[4..]
         else displayName
 
-    let internal isEventMember (m: FSharpMemberFunctionOrValue) =
+    let internal isEventMember (m: FSharpMemberOrFunctionOrValue) =
         m.IsEvent || hasAttribute<CLIEventAttribute> m.Attributes
     
     let internal formatMember (ctx: Context) m = 
-        let getParamArgs (argInfos: FSharpParameter list list) (ctx: Context) (v: FSharpMemberFunctionOrValue) = 
+        let getParamArgs (argInfos: FSharpParameter list list) (ctx: Context) (v: FSharpMemberOrFunctionOrValue) = 
             let args, namesWithIndices =
                 match argInfos with
                 | [[x]] when v.IsPropertyGetterMethod && x.Name.IsNone 
@@ -210,7 +210,7 @@ module InterfaceStubGenerator =
             else sprintf "(%s)" args
             , namesWithIndices
 
-        let preprocess (ctx: Context) (v: FSharpMemberFunctionOrValue) = 
+        let preprocess (ctx: Context) (v: FSharpMemberOrFunctionOrValue) = 
             let buildUsage argInfos = 
                 let parArgs, _ = getParamArgs argInfos ctx v
                 match v.IsMember, v.IsInstanceMember, v.LogicalName, v.DisplayName with
@@ -403,7 +403,7 @@ module InterfaceStubGenerator =
             Some typ.GenericArguments.[1]
         else None
 
-    let internal (|TypeOfMember|_|) (m: FSharpMemberFunctionOrValue) =
+    let internal (|TypeOfMember|_|) (m: FSharpMemberOrFunctionOrValue) =
         match m.FullTypeSafe with
         | Some (MemberFunctionType typ) when m.IsProperty && m.EnclosingEntity.IsFSharp ->
             Some typ
@@ -426,7 +426,7 @@ module InterfaceStubGenerator =
     let internal removeWhitespace (str: string) = 
         str.Replace(" ", "")
 
-    let internal normalizeEventName (m: FSharpMemberFunctionOrValue) =
+    let internal normalizeEventName (m: FSharpMemberOrFunctionOrValue) =
         let name = m.DisplayName
         if name.StartsWith("add_") then name.[4..]
         elif name.StartsWith("remove_")  then name.[7..]
@@ -440,7 +440,7 @@ module InterfaceStubGenerator =
     let getImplementedMemberSignatures (getMemberByLocation: string * range -> Async<FSharpSymbolUse option>) displayContext interfaceData = 
         let formatMemberSignature (symbolUse: FSharpSymbolUse) =            
             match symbolUse.Symbol with
-            | :? FSharpMemberFunctionOrValue as m ->
+            | :? FSharpMemberOrFunctionOrValue as m ->
                 match m.FullTypeSafe with
                 | Some _ when isEventMember m ->
                     // Events don't have overloads so we use only display names for comparison
@@ -511,7 +511,7 @@ module InterfaceStubGenerator =
             writer.Indent startColumn
 
             let getReturnType v = snd (getArgTypes ctx v)
-            let rec loop (members : (FSharpMemberFunctionOrValue * _) list) =
+            let rec loop (members : (FSharpMemberOrFunctionOrValue * _) list) =
                 match members with
                 // Since there is no unified source of information for properties,
                 // we try to merge getters and setters when they seem to match.