+++ tests/FSharpVSPowerTools.Core.Tests/UnionPatternMatchCaseGeneratorTests.fs
@@ -1,6 +1,6 @@
 ﻿#if INTERACTIVE
 #r "../../bin/FSharp.Compiler.Service.dll"
-#r "../../packages/NUnit.2.6.3/lib/nunit.framework.dll"
+#r "../../packages/NUnit/lib/nunit.framework.dll"
 #load "../../src/FSharpVSPowerTools.Core/Utils.fs"
       "../../src/FSharpVSPowerTools.Core/CompilerLocationUtils.fs"
       "../../src/FSharpVSPowerTools.Core/Lexer.fs"
@@ -32,7 +32,7 @@ let project() =
     let projFileName = @"C:\Project.fsproj"
     { ProjectFileName = projFileName
       ProjectFileNames = [| fileName |]
-      ProjectOptions = LanguageServiceTestHelper.args
+      OtherOptions = LanguageServiceTestHelper.args
       ReferencedProjects = Array.empty
       IsIncompleteTypeCheckEnvironment = false
       UseScriptResolutionRules = false